+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/AwsLambdaDetector.ts
@@ -21,9 +21,13 @@ import {
   ResourceDetectionConfig,
 } from '@opentelemetry/resources';
 import {
-  CloudProviderValues,
-  CloudPlatformValues,
-  SemanticResourceAttributes,
+  SEMRESATTRS_CLOUD_PROVIDER,
+  SEMRESATTRS_CLOUD_PLATFORM,
+  SEMRESATTRS_CLOUD_REGION,
+  SEMRESATTRS_FAAS_VERSION,
+  SEMRESATTRS_FAAS_NAME,
+  CLOUDPROVIDERVALUES_AWS,
+  CLOUDPLATFORMVALUES_AWS_LAMBDA,
 } from '@opentelemetry/semantic-conventions';
 
 /**
@@ -42,22 +46,22 @@ export class AwsLambdaDetector implements Detector {
     const region = process.env.AWS_REGION;
 
     const attributes: ResourceAttributes = {
-      [SemanticResourceAttributes.CLOUD_PROVIDER]: String(
-        CloudProviderValues.AWS
+      [SEMRESATTRS_CLOUD_PROVIDER]: String(
+        CLOUDPROVIDERVALUES_AWS
       ),
-      [SemanticResourceAttributes.CLOUD_PLATFORM]: String(
-        CloudPlatformValues.AWS_LAMBDA
+      [SEMRESATTRS_CLOUD_PLATFORM]: String(
+        CLOUDPLATFORMVALUES_AWS_LAMBDA
       ),
     };
     if (region) {
-      attributes[SemanticResourceAttributes.CLOUD_REGION] = region;
+      attributes[SEMRESATTRS_CLOUD_REGION] = region;
     }
 
     if (functionName) {
-      attributes[SemanticResourceAttributes.FAAS_NAME] = functionName;
+      attributes[SEMRESATTRS_FAAS_NAME] = functionName;
     }
     if (functionVersion) {
-      attributes[SemanticResourceAttributes.FAAS_VERSION] = functionVersion;
+      attributes[SEMRESATTRS_FAAS_VERSION] = functionVersion;
     }
 
     return new Resource(attributes);