+++ detectors/node/opentelemetry-resource-detector-azure/test/detectors/AzureAppServiceDetector.test.ts
@@ -16,7 +16,15 @@
 
 import * as assert from 'assert';
 import { azureAppServiceDetector } from '../../src/detectors/AzureAppServiceDetector';
-import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions';
+import { 
+  SEMRESATTRS_CLOUD_PLATFORM,
+  SEMRESATTRS_CLOUD_PROVIDER,
+  SEMRESATTRS_CLOUD_REGION,
+  SEMRESATTRS_DEPLOYMENT_ENVIRONMENT,
+  SEMRESATTRS_HOST_ID,
+  SEMRESATTRS_SERVICE_INSTANCE_ID,
+  SEMRESATTRS_SERVICE_NAME
+} from '@opentelemetry/semantic-conventions';
 import { azureFunctionsDetector } from '../../src';
 import { detectResourcesSync } from '@opentelemetry/resources';
 
@@ -46,15 +54,15 @@ describe('AzureAppServiceDetector', () => {
     assert.ok(resource);
     const attributes = resource.attributes;
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.SERVICE_NAME],
+      attributes[SEMRESATTRS_SERVICE_NAME],
       'test-site'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.CLOUD_PROVIDER],
+      attributes[SEMRESATTRS_CLOUD_PROVIDER],
       'azure'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.CLOUD_PLATFORM],
+      attributes[SEMRESATTRS_CLOUD_PLATFORM],
       'azure_app_service'
     );
     assert.strictEqual(
@@ -62,19 +70,19 @@ describe('AzureAppServiceDetector', () => {
       `/subscriptions/${process.env.WEBSITE_OWNER_NAME}/resourceGroups/${process.env.WEBSITE_RESOURCE_GROUP}/providers/Microsoft.Web/sites/${process.env.WEBSITE_SITE_NAME}`
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.CLOUD_REGION],
+      attributes[SEMRESATTRS_CLOUD_REGION],
       'test-region'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.DEPLOYMENT_ENVIRONMENT],
+      attributes[SEMRESATTRS_DEPLOYMENT_ENVIRONMENT],
       'test-slot'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.HOST_ID],
+      attributes[SEMRESATTRS_HOST_ID],
       'test-hostname'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.SERVICE_INSTANCE_ID],
+      attributes[SEMRESATTRS_SERVICE_INSTANCE_ID],
       'test-instance-id'
     );
     assert.strictEqual(
@@ -98,19 +106,19 @@ describe('AzureAppServiceDetector', () => {
     assert.ok(resource);
     const attributes = resource.attributes;
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.CLOUD_REGION],
+      attributes[SEMRESATTRS_CLOUD_REGION],
       'test-region'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.DEPLOYMENT_ENVIRONMENT],
+      attributes[SEMRESATTRS_DEPLOYMENT_ENVIRONMENT],
       'test-slot'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.HOST_ID],
+      attributes[SEMRESATTRS_HOST_ID],
       'test-hostname'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.SERVICE_INSTANCE_ID],
+      attributes[SEMRESATTRS_SERVICE_INSTANCE_ID],
       'test-instance-id'
     );
     assert.strictEqual(
@@ -135,19 +143,19 @@ describe('AzureAppServiceDetector', () => {
     assert.ok(resource);
     const attributes = resource.attributes;
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.CLOUD_REGION],
+      attributes[SEMRESATTRS_CLOUD_REGION],
       'test-region'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.DEPLOYMENT_ENVIRONMENT],
+      attributes[SEMRESATTRS_DEPLOYMENT_ENVIRONMENT],
       'test-slot'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.HOST_ID],
+      attributes[SEMRESATTRS_HOST_ID],
       'test-hostname'
     );
     assert.strictEqual(
-      attributes[SemanticResourceAttributes.SERVICE_INSTANCE_ID],
+      attributes[SEMRESATTRS_SERVICE_INSTANCE_ID],
       'test-instance-id'
     );
     assert.strictEqual(