+++ examples/connect/tracing.js
@@ -24,7 +24,7 @@ function log() {
 module.exports = (serviceName) => {
   const provider = new NodeTracerProvider({
     resource: new Resource({
-      [SemanticResourceAttributes.SERVICE_NAME]: serviceName,
+      [SEMRESATTRS_SERVICE_NAME]: serviceName,
     }),
   });
   const connectInstrumentation = new ConnectInstrumentation();